// src/protos/InteractWordV2.ts
export interface InteractWordV2 {
    /**
     * @generated from protobuf field: uint64 uid = 1
     */
    uid: bigint;
    /**
     * @generated from protobuf field: string uname = 2
     */
    uname: string;
    /**
     * @generated from protobuf field: InteractWordV2.MsgType msg_type = 5
     */
    msgType: InteractWordV2_MsgType;
    /**
     * @generated from protobuf field: uint32 room_id = 6
     */
    roomId: number;
    /**
     * @generated from protobuf field: uint32 timestamp = 7
     */
    timestamp: number; // second
    /**
     * @generated from protobuf field: uint64 score = 8
     */
    score: bigint;
    /**
     * @generated from protobuf field: optional InteractWordV2.FansMedal fans_medal = 9
     */
    fansMedal?: InteractWordV2_FansMedal;
    /**
     * @generated from protobuf field: InteractWordV2.RankInfo rank_info = 12
     */
    rankInfo?: InteractWordV2_RankInfo;
    /**
     * @generated from protobuf field: uint64 trigger_time = 15
     */
    triggerTime: bigint; // maybe
    /**
     * @generated from protobuf field: optional uint32 guard_level = 16
     */
    guardLevel?: number; // 可能对应INTERACT_WORD中的privilege_type
    /**
     * @generated from protobuf field: optional uint32 daily_rank = 17
     */
    dailyRank?: number;
    /**
     * @generated from protobuf field: InteractWordV2.UserInfo uinfo = 22
     */
    uinfo?: InteractWordV2_UserInfo;
}
/**
 * @generated from protobuf message InteractWordV2.FansMedal
 */
export interface InteractWordV2_FansMedal {
    /**
     * @generated from protobuf field: uint64 target_id = 1
     */
    targetId: bigint; // also named anchor_id or ruid
    /**
     * @generated from protobuf field: uint32 medal_level = 2
     */
    medalLevel: number;
    /**
     * @generated from protobuf field: string medal_name = 3
     */
    medalName: string;
    /**
     * @generated from protobuf field: uint32 color = 4
     */
    color: number;
    /**
     * @generated from protobuf field: uint32 color_start = 5
     */
    colorStart: number;
    /**
     * @generated from protobuf field: uint32 color_end = 6
     */
    colorEnd: number;
    /**
     * @generated from protobuf field: uint32 color_border = 7
     */
    colorBorder: number;
    /**
     * @generated from protobuf field: optional uint32 is_lighted = 8
     */
    isLighted?: number;
    /**
     * @generated from protobuf field: optional uint32 guard_level = 9
     */
    guardLevel?: number;
    /**
     * @generated from protobuf field: uint32 room_id = 12
     */
    roomId: number;
    /**
     * @generated from protobuf field: uint32 score = 13
     */
    score: number;
}
/**
 * @generated from protobuf message InteractWordV2.RankInfo
 */
export interface InteractWordV2_RankInfo {
    /**
     * @generated from protobuf field: uint32 rank = 1
     */
    rank: number;
}
// 19: 排行榜信息
// message DailyRankInfo {
//     uint32 rank = 1;
//     string key = 2;
//     string description = 3;
// }
// optional DaylyRankInfo daily_rank_info = 19;

/**
 * @generated from protobuf message InteractWordV2.UserInfo
 */
export interface InteractWordV2_UserInfo {
    /**
     * @generated from protobuf field: uint64 uid = 1
     */
    uid: bigint;
    /**
     * @generated from protobuf field: InteractWordV2.UserInfo.BaseInfo base = 2
     */
    base?: InteractWordV2_UserInfo_BaseInfo;
    /**
     * @generated from protobuf field: optional InteractWordV2.UserInfo.MedalInfo medal_info = 3
     */
    medalInfo?: InteractWordV2_UserInfo_MedalInfo;
    /**
     * @generated from protobuf field: optional InteractWordV2.UserInfo.WealthInfo wealth = 4
     */
    wealth?: InteractWordV2_UserInfo_WealthInfo;
    /**
     * @generated from protobuf field: optional InteractWordV2.UserInfo.GuardInfo guard = 6
     */
    guard?: InteractWordV2_UserInfo_GuardInfo;
}
/**
 * @generated from protobuf message InteractWordV2.UserInfo.BaseInfo
 */
export interface InteractWordV2_UserInfo_BaseInfo {
    /**
     * @generated from protobuf field: string uname = 1
     */
    uname: string;
    /**
     * @generated from protobuf field: string face = 2
     */
    face: string;
    /**
     * @generated from protobuf field: optional InteractWordV2.UserInfo.BaseInfo.OriginInfo origin_info = 6
     */
    originInfo?: InteractWordV2_UserInfo_BaseInfo_OriginInfo;
    /**
     * message OnlyGodKnow_1 {
     *    uint64 i_dont_know = 4;
     * }
     * optional OnlyGodKnow_1 only_god_know = 7;
     *
     * @generated from protobuf field: optional string name_color_str = 8
     */
    nameColorStr?: string; // username color showed in danmaku
}
// ...
export interface InteractWordV2_UserInfo_MedalInfo {
    /**
     * @generated from protobuf field: string medal_name = 1
     */
    medalName: string;
    /**
     * @generated from protobuf field: uint32 medal_level = 2
     */
    medalLevel: number;
    /**
     * @generated from protobuf field: uint32 color_start = 3
     */
    colorStart: number;
    /**
     * @generated from protobuf field: uint32 color_end = 4
     */
    colorEnd: number;
    /**
     * @generated from protobuf field: uint32 color_border = 5
     */
    colorBorder: number;
    /**
     * @generated from protobuf field: uint32 color = 6
     */
    color: number;
    /**
     * @generated from protobuf field: uint32 id = 7
     */
    id: number; // I don't know what is this
    /**
     * @generated from protobuf field: uint32 is_lighted = 9
     */
    isLighted: number;
    /**
     * @generated from protobuf field: uint64 ruid = 10
     */
    ruid: bigint; // also named target_id or anchor_id
    /**
     * @generated from protobuf field: uint32 guard_level = 11
     */
    guardLevel: number;
    /**
     * @generated from protobuf field: uint32 score = 12
     */
    score: number;
    /**
     * @generated from protobuf field: string guard_icon = 13
     */
    guardIcon: string;
    /**
     * @generated from protobuf field: string v2_medal_color_start = 15
     */
    v2MedalColorStart: string;
    /**
     * @generated from protobuf field: string v2_medal_color_end = 16
     */
    v2MedalColorEnd: string;
    /**
     * @generated from protobuf field: string v2_medal_color_border = 17
     */
    v2MedalColorBorder: string;
    /**
     * @generated from protobuf field: string v2_medal_color_text = 18
     */
    v2MedalColorText: string;
    /**
     * @generated from protobuf field: string v2_medal_color_level = 19
     */
    v2MedalColorLevel: string;
}
/**
 * @generated from protobuf message InteractWordV2.UserInfo.WealthInfo
 */
export interface InteractWordV2_UserInfo_WealthInfo {
    /**
     * @generated from protobuf field: uint32 level = 1
     */
    level: number;
}
/**
 * @generated from protobuf message InteractWordV2.UserInfo.GuardInfo
 */
export interface InteractWordV2_UserInfo_GuardInfo {
    /**
     * @generated from protobuf field: uint32 level = 1
     */
    level: number;
    /**
     * @generated from protobuf field: string expired_str = 2
     */
    expiredStr: string;
}
/**
 * @generated from protobuf enum InteractWordV2.MsgType
 */
export enum InteractWordV2_MsgType {
    /**
     * @generated from protobuf enum value: MSG_UNKNOWN = 0;
     */
    MSG_UNKNOWN = 0,
    /**
     * @generated from protobuf enum value: MSG_ENTER_ROOM = 1;
     */
    MSG_ENTER_ROOM = 1,
    /**
     * @generated from protobuf enum value: MSG_FOLLOW = 2;
     */
    MSG_FOLLOW = 2,
    /**
     * @generated from protobuf enum value: MSG_SHARE_ROOM = 3;
     */
    MSG_SHARE_ROOM = 3
}
// ...
export const InteractWordV2 = new InteractWordV2$Type();
// ...
export const InteractWordV2_FansMedal = new InteractWordV2_FansMedal$Type();
// @generated message type with reflection information, may provide speed optimized methods
class InteractWordV2_RankInfo$Type extends MessageType<InteractWordV2_RankInfo> {
    constructor() {
        super("InteractWordV2.RankInfo", [
            { no: 1, name: "rank", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<InteractWordV2_RankInfo>): InteractWordV2_RankInfo {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.rank = 0;
        if (value !== undefined)
            reflectionMergePartial<InteractWordV2_RankInfo>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: InteractWordV2_RankInfo): InteractWordV2_RankInfo {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint32 rank */ 1:
                    message.rank = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: InteractWordV2_RankInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint32 rank = 1; */
        if (message.rank !== 0)
            writer.tag(1, WireType.Varint).uint32(message.rank);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message InteractWordV2.RankInfo
 */
export const InteractWordV2_RankInfo = new InteractWordV2_RankInfo$Type();
// ...
export const InteractWordV2_UserInfo = new InteractWordV2_UserInfo$Type();
// ...
export const InteractWordV2_UserInfo_BaseInfo = new InteractWordV2_UserInfo_BaseInfo$Type();
// ...
export const InteractWordV2_UserInfo_MedalInfo = new InteractWordV2_UserInfo_MedalInfo$Type();
// ...
export const InteractWordV2_UserInfo_WealthInfo = new InteractWordV2_UserInfo_WealthInfo$Type();
// ...
export const InteractWordV2_UserInfo_GuardInfo = new InteractWordV2_UserInfo_GuardInfo$Type();